mac springboot qna step5.4-- 질문 수정/삭제 기능에대한 보안처리

+++ src/main/java/com/nekisse/com/qna/web/QuestionController.java
@@ -39,8 +39,6 @@ public class QuestionController {
         User userSession = HttpSessionUtils.getUserFromSession(session);
         Question newQuestion = new Question(userSession, title, contents);
         questionRepository.save(newQuestion);
-
-
         return "redirect:/";
     }
 
@@ -51,23 +49,53 @@ public class QuestionController {
         return "qna/show";
     }
 
-    @PutMapping("/{id}")
-    public String update(@PathVariable Long id, String title, String contents) {
+    @GetMapping("{id}/form")
+    public String updateForm(@PathVariable Long id, Model model, HttpSession session) {
+        if (!HttpSessionUtils.isLoginUser(session)) {
+            return "/users/loginForm";
+        }
+
+        User loginUser = HttpSessionUtils.getUserFromSession(session);
         Question question = questionRepository.getOne(id);
+        if (!question.isSameWriter(loginUser)) {
+            return "users/loginForm";
+        }
+
+        model.addAttribute("question", question);
+        return "qna/updateForm";
+
+    }
+
+    @PutMapping("/{id}")
+    public String update(@PathVariable Long id, String title, String contents, HttpSession session) {
+        if (!HttpSessionUtils.isLoginUser(session)) {
+            return "/users/loginForm";
+        }
+
+        User loginUser = HttpSessionUtils.getUserFromSession(session);
+        Question question = questionRepository.getOne(id);
+        if (!question.isSameWriter(loginUser)) {
+            return "users/loginForm";
+        }
+
         question.update(title, contents);
         questionRepository.save(question);
         return String.format("redirect:/questions/%d", id);
     }
 
     @DeleteMapping("/{id}")
-    public String delete(@PathVariable Long id) {
+    public String delete(@PathVariable Long id, HttpSession session) {
+        if (!HttpSessionUtils.isLoginUser(session)) {
+            return "/users/loginForm";
+        }
+
+        User loginUser = HttpSessionUtils.getUserFromSession(session);
+        Question question = questionRepository.getOne(id);
+        if (!question.isSameWriter(loginUser)) {
+            return "users/loginForm";
+        }
+
         questionRepository.deleteById(id);
         return "redirect:/";
-    }
-    @GetMapping("{id}/form")
-    public String updateForm(@PathVariable Long id, Model model) {
-        model.addAttribute("question", questionRepository.getOne(id));
-        return "qna/updateForm";
-
     }
 }
